* commit '1e1feb0614c81658a001814c7b28f5b24d33c4d6': (47 commits)
  Cleanup. Fix test.
  Clean up parameters classes.
  Make BulletContactSolverInfoParameters constants doubles. Potentially changes time step from 0 to 1.0 / 60.0.
  Remove old comments about debug drawer.
  Remove old libGDX commented method.
  Extract BulletBroadphaseNativeTypes, map to the native values, and fix a few shape casts. Getting AssertionError.
  Uncomment and fix shape casting. A little cleanup.
  Add getRotation to test. Shouldn't change functionality, just to be clear.
  Clean up Bullet debug drawing node and fix warnings.
  Remove snapshots declaration.
  Remove snapshots declaration.
  Add documentation for coefficient tests.
  Found the solver contants.
  Track terrain object friction. Allow setting terrain friction in tests. Fix coeficient of friction test and add test cases.
  Shutdown sessions to fix deallocator native crashes.
  Formatting and adding dot zeros.
  Fix native deallocator crashes by shutting down the sessions.
  Fix test method alignment.
  Fix Bullet terrain factory test with the Bytedeco bindings.
  Fix multi body parameters test.
  ...

+++ scs2-bullet-simulation/src/main/java/us/ihmc/scs2/simulation/bullet/physicsEngine/BulletRobotLinkJoint.java
@@ -51,9 +51,9 @@ public class BulletRobotLinkJoint extends BulletRobotLinkBasics
 
    public void pullStateFromBullet(double dt)
    {
-      float jointPosition = getBulletMultiBodyLinkCollider().getJointPos();
+      double jointPosition = getBulletMultiBodyLinkCollider().getJointPos();
       simOneDofJoint.setQ(jointPosition);
-      float jointPVel = getBulletMultiBodyLinkCollider().getJointVel();
+      double jointPVel = getBulletMultiBodyLinkCollider().getJointVel();
       simOneDofJoint.setQdd((jointPVel - simOneDofJoint.getQd()) / dt);
       simOneDofJoint.setQd(jointPVel);
